requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.4
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
importlib-metadata==2.1.1
Pillow==8.2.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
